Added contact form to the frontpage of the site (#5)
* Starting portfolio 
A portfolio to show off my projects
Just throwing in the empty files needed and the document heirarchy notes 
for the HTML.

* spelling mistakes fixed

* Moved Portfolio.js to correct folder 
portfolio.js was created one level below where it should be in in the 
file structure. It is now in the correct location.

* Adding images to folder
mongoDB.svg is an icon for mongoDB
placeholder.svg is the default 'put image here' icon for the page.

* Fixed skill_icon height. 
Skill icons were rendering at various heights, all of them wrong.
Pinned skill_icons to a set height of 1.25rem.

* First pass of the content.
With minimal CSS and almost no formatting the first pass bad, but 
passable on mobile devices and desktops. It will basicly run on a 
potato.

* Moved portfolio files to correct folder. 
Portfolio was in projects folder, moved to it's own folder one level up.

* set project_image width to 99vw

* Created a notes.txt file

* Initial CSS for mobile devices.

* Added viewport meta info and linked to main stylesheet for site

* Initial .js for page.
The closed tag is now added to all Project_tiles upon loading the 
document.
Even listener added to all project_tiles: they will now open/close when 
clicked.

* Smoothed project_tile autoscrolling. 
Added EventListener to all tiles, they now wait until the transition has 
ended to scroll into view.

* Closed tiles now have a blur effect applied to the project_image

* bug-fix: eventlistener checking for wrong property 
Event listener was cheing for transion on 'width', should be 
'max-width'. Fixed.

* removed debugging console.log()

* bugfix: Removed repeated element id.
The This Portfolio tile was sharring the id #portfolio with the 
portfolio container element, changed This Portfolio's id to 
#thisPortfolio

* project_tiles open/close now works for both height and width.

* First breakpoint 340px working.
Above 340px wide, the page switches to from opening tiles in a vertical 
format to a horizontal format

* CSS Specific to portfolio ready to go.

* cursur now pointing at clickable elements
project_miniview and all the project links now have 'cursor: pointer' in 
their CSS.
Added '.noPointy' class to CSS for elements that need the pointer 
suppressed.

* Updated to the new color pallet.
The volor vars are now set to a new pallet that was pulled from the 
banner image of the index-page.

* Added new background image and credits for image.
Background image pulled from svgimages.com

* Added faded background image to .col

* Cleaned up the borders of the .col and .colImg

* fixed typos in URLs

* Updated images for the resume, portfolio and contact tiles.

* fixed url for card catalog view.

* Added background to .project_tiles.
Gave .project_tiles the same subtly textured background as the tiles in 
the frontpage.

* fixed bugged data url for the .project_tile background image.

* decrease opacity of the tile background texture.

* Added images for card Catalog FrontEnd and BackEnd tiles.

* Fixed blur on image bleedig outside border of image.

* Updated image for #thisportfolio

* Updated image for #messageboard

* fixed color of mongoDb.svg to match theme of rest of site.

* Added nav-bar to portfolio.

* Form and fields added to index.html

* Fixed submit button value.
Submit button was not appearing becasue no value property was set for 
it, Set value='Submit', working now.

* Contact form is added, styled and working.
Contact form has been styled to match look of the rest of the sight.
The 'contact' link now smooth-scrolls to the contact form.
Contact form POSTs to formspree.io which then sends the information to 
me as an email.